requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.6
cycler==0.10.0
idna==3.2
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
panda==0.3.1
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.7
